Added .gitmessage template for commit messages
Added .gitmessage template to guide commit messages. This template
includes instructions on how to write a good commit message,
including the structure of the title and body, and how to format
co-authored commits. It also provides a link to a resource for
writing good commit messages. This will help maintain a consistent
commit message style across the project, making it easier to
understand the history of changes and the reasons behind them.
https://gist.github.com/lisawolderiksen/a7b99d94c92c6671181611be1641c733